July 22 Added more to Report class. The push total sales method works now and the methods have been created to be used to update the database with all the flavour sales.
Next Sale class. This will open a new window for sales and once the sale is confirmed if will first test to ensure that there is enough stock, this is tested with the stockList (or maybe better to test on adding to the list) next it  will execute the report writing writing a report for the file and exporting it to the database.